pyproject.toml
[tool.poetry]
name = "PriathiyIldar"
version = "0.1.0"
description = ""
authors = ["Author <author@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
drf-spectacular = "^0.26.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-django = "^1.4"
flake8-pyproject = "^1.2.3"
mypy = "^1.7.0"
black = "^23.11.0"
isort = "^5.12.0"
coverage = "^7.3.2"
types-requests = "^2.31.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
count = true
ignore = ['E203']

[tool.mypy]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
    | \.mypy_cache
  )/
)
'''
